requirements.txt
telethon==1.36.0
easyocr==1.7.2
python-dotenv==1.0.1
Pillow==11.1.0
